requirements.txt
requests
python-dotenv
fastapi
markdownify
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client